+++ NEWS.md
@@ -1,6 +1,8 @@
 # fastRG (development version)
 
+* Add degree-corrected overlapping stochastic blockmodel (#7)
 * Add directed degree-corrected stochastic blockmodels (#18)
+* Minor bug fixes for SBM-related functions
 
 # fastRG 0.3.0